fix(config): add setupConfig util
