--- core/src/global/queue-controller.ts
@@ -1,13 +1,0 @@
-
-export interface QueueController {
-  read: DomControllerCallback;
-  write: DomControllerCallback;
-}
-
-export interface RafCallback {
-  (timeStamp: number): void;
-}
-
-export interface DomControllerCallback {
-  (cb: RafCallback): void;
-}
